mv plots to utils
